chore: rename lint job to Swagger Decorator Lint and update Node.js version to 20
