+++ backend/rules/require-api-response.cjs
@@ -1,9 +1,9 @@
-// rules/require-api-produces.mjs
-export default {
+// rules/require-api-response.mjs
+module.exports =  {
     meta: {
       type: 'problem',
       docs: {
-        description: 'Ensure every controller method has an @ApiProduces() decorator',
+        description: 'Ensure every controller method has at least one @ApiResponse() decorator',
       },
       schema: [],
     },
@@ -15,8 +15,8 @@ export default {
   
       return {
         ClassDeclaration(node) {
-          const decorators = node.decorators || [];
-          const isController = decorators.some((dec) => {
+          // Only inspect @Controller classes
+          const isController = (node.decorators || []).some((dec) => {
             const expr = dec.expression;
             return (
               expr.type === 'CallExpression' &&
@@ -27,7 +27,6 @@ export default {
           if (!isController) return;
   
           for (const member of node.body.body) {
-            // only real methods (skip constructors, getters, setters, etc.)
             if (
               member.type !== 'MethodDefinition' ||
               member.kind !== 'method' ||
@@ -36,19 +35,19 @@ export default {
               continue;
             }
   
-            const hasProduces = (member.decorators || []).some((dec) => {
+            const hasApiResponse = (member.decorators || []).some((dec) => {
               const expr = dec.expression;
               return (
                 expr.type === 'CallExpression' &&
                 expr.callee.type === 'Identifier' &&
-                expr.callee.name === 'ApiProduces'
+                expr.callee.name === 'ApiResponse'
               );
             });
   
-            if (!hasProduces) {
+            if (!hasApiResponse) {
               context.report({
                 node: member.key,
-                message: `Method '${member.key.name}' is missing @ApiProduces() decorator.`,
+                message: `Method '${member.key.name}' is missing at least one @ApiResponse() decorator.`,
               });
             }
           }
